Move SortNode to spi so that it could be used in connector
