+++ presto-main/src/main/java/com/facebook/presto/cost/SortStatsRule.java
@@ -16,9 +16,9 @@ package com.facebook.presto.cost;
 import com.facebook.presto.Session;
 import com.facebook.presto.cost.ComposableStatsCalculator.Rule;
 import com.facebook.presto.matching.Pattern;
+import com.facebook.presto.spi.plan.SortNode;
 import com.facebook.presto.sql.planner.TypeProvider;
 import com.facebook.presto.sql.planner.iterative.Lookup;
-import com.facebook.presto.sql.planner.plan.SortNode;
 
 import java.util.Optional;
