+++ presto-main/src/main/java/com/facebook/presto/sql/planner/iterative/rule/SimplifySortWithConstantInput.java
@@ -20,10 +20,10 @@ import com.facebook.presto.matching.Pattern;
 import com.facebook.presto.spi.plan.Ordering;
 import com.facebook.presto.spi.plan.OrderingScheme;
 import com.facebook.presto.spi.plan.ProjectNode;
+import com.facebook.presto.spi.plan.SortNode;
 import com.facebook.presto.spi.relation.ConstantExpression;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
 import com.facebook.presto.sql.planner.iterative.Rule;
-import com.facebook.presto.sql.planner.plan.SortNode;
 
 import java.util.List;
 import java.util.Map;
