+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -72,11 +72,6 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitSort(SortNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitWindow(WindowNode node, C context)
     {
         return visitPlan(node, context);
